requirements.txt
discord.py

# Date and Time Handling
python-dateutil

# Settings Handling
python-dotenv
pydantic-settings

# Files
aiofiles
